Update daysUntilBirthday.js
As in my last probosel, there is an interface to select the adressbooks. The selection is stored.
In this version a added an option to show family names or not. If family names are shown the widget only has one column. Without family name it is the old design with two columns.
I switched back to the old colors but if the birthday is today, the color will be red.